+++ Assets/OxGFrame/CoreFrame/Example/EPFrameDemo/Scripts/DemoEntity1.cs
@@ -3,15 +3,13 @@ using UnityEngine;
 
 public class DemoEntity1 : EPBase
 {
-    public override void BeginInit()
+    public override void OnInit()
     {
         Debug.Log($"<color=#FF2A20>InitThis:</color> {this.gameObject.name}");
     }
 
-    protected override void InitOnceComponents()
+    protected override void OnBind()
     {
-        Debug.Log($"<color=#FFA720>InitOnceComponents:</color> {this.gameObject.name}");
-
         // Single Bind
         Debug.Log($"<color=#FFA720>Found:</color> {this.gameObject.name} => {this.collector.GetNode("B1").name}");
 
@@ -21,11 +19,6 @@ public class DemoEntity1 : EPBase
         {
             Debug.Log($"<color=#FFA720>Found:</color> {this.gameObject.name} => {node.name}");
         }
-    }
-
-    protected override void InitOnceEvents()
-    {
-        Debug.Log($"<color=#FFE920>InitOnceEvents:</color> {this.gameObject.name}");
     }
 
     protected override void OnShow(object obj)