+++ Assets/OxGFrame/CoreFrame/Example/GSFrameDemo/Scripts/DemoSC.cs
@@ -6,7 +6,7 @@ using OxGFrame.CoreFrame.GSFrame;
 
 public class DemoSC : GSBase
 {
-    public override void BeginInit()
+    public override void OnInit()
     {
     }
 
@@ -29,18 +29,11 @@ public class DemoSC : GSBase
         Debug.Log("DemoSC OnShow");
     }
 
-    protected override void InitOnceComponents()
+    protected override void OnBind()
     {
         /**
-         * Do Somthing Init Once In Here (For Components)
+         * Do Somthing Init Once Components and Events In Here
          */
-    }
-
-    protected override void InitOnceEvents()
-    {
-        /**
-          * Do Somthing Init Once In Here (For Events)
-          */
     }
 
     protected override void OnUpdate(float dt)