+++ Assets/OxGFrame/CoreFrame/Example/UIFrameDemo/Scripts/Demo1UI.cs
@@ -10,7 +10,7 @@ public class Demo1UI : UIBase
     private Button oepnBtn;
     private Image myImage2;
 
-    public override void BeginInit()
+    public override void OnInit()
     {
     }
 
@@ -33,17 +33,14 @@ public class Demo1UI : UIBase
         Debug.Log(string.Format("{0} Do Something OnShow.", this.gameObject.name));
     }
 
-    protected override void InitOnceComponents()
+    protected override void OnBind()
     {
         this.myImage = this.collector.GetNode("Image1")?.GetComponent<Image>();
         if (this.myImage != null) Debug.Log(string.Format("Binded GameObject: {0}", this.myImage.name));
 
         this.oepnBtn = this.collector.GetNode("OpenBtn")?.GetComponent<Button>();
         if (this.oepnBtn != null) Debug.Log(string.Format("Binded GameObject: {0}", this.oepnBtn.name));
-    }
 
-    protected override void InitOnceEvents()
-    {
         this.oepnBtn.onClick.AddListener(this._ShowDemoPopup2UI);
     }